Move eip191 tests to lcp
